GUI: Moved GeneralServiceManager API usage to ServicesManager
- Moved classes and API usage from GeneralServiceManager to
  ServicesManager.
