+++ perun-web-gui/src/main/java/cz/metacentrum/perun/webgui/tabs/facilitiestabs/DestinationResultsTabItem.java
@@ -16,8 +16,8 @@ import cz.metacentrum.perun.webgui.client.resources.TableSorter;
 import cz.metacentrum.perun.webgui.json.GetEntityById;
 import cz.metacentrum.perun.webgui.json.JsonCallbackEvents;
 import cz.metacentrum.perun.webgui.json.JsonUtils;
-import cz.metacentrum.perun.webgui.json.generalServiceManager.ForceServicePropagation;
-import cz.metacentrum.perun.webgui.json.generalServiceManager.GetFacilityAssignedServicesForGUI;
+import cz.metacentrum.perun.webgui.json.servicesManager.ForceServicePropagation;
+import cz.metacentrum.perun.webgui.json.servicesManager.GetFacilityAssignedServicesForGUI;
 import cz.metacentrum.perun.webgui.json.propagationStatsReader.GetTaskResultsByDestinations;
 import cz.metacentrum.perun.webgui.model.*;
 import cz.metacentrum.perun.webgui.tabs.*;
