+++ perun-web-gui/src/main/java/cz/metacentrum/perun/webgui/tabs/facilitiestabs/FacilityStatusTabItem.java
@@ -15,9 +15,9 @@ import cz.metacentrum.perun.webgui.client.resources.*;
 import cz.metacentrum.perun.webgui.json.GetEntityById;
 import cz.metacentrum.perun.webgui.json.JsonCallbackEvents;
 import cz.metacentrum.perun.webgui.json.JsonUtils;
-import cz.metacentrum.perun.webgui.json.generalServiceManager.BlockServiceOnFacility;
-import cz.metacentrum.perun.webgui.json.generalServiceManager.ForceServicePropagation;
-import cz.metacentrum.perun.webgui.json.generalServiceManager.UnblockServiceOnFacility;
+import cz.metacentrum.perun.webgui.json.servicesManager.BlockServiceOnFacility;
+import cz.metacentrum.perun.webgui.json.servicesManager.ForceServicePropagation;
+import cz.metacentrum.perun.webgui.json.servicesManager.UnblockServiceOnFacility;
 import cz.metacentrum.perun.webgui.json.propagationStatsReader.DeleteTask;
 import cz.metacentrum.perun.webgui.json.propagationStatsReader.GetFacilityServicesState;
 import cz.metacentrum.perun.webgui.model.*;
